Issue-3499: Move parallel module to compose/core package
Partially implements #3499
