+++ compose/core/parallel.py
@@ -10,8 +10,8 @@ from six.moves import _thread as thread
 from six.moves.queue import Empty
 from six.moves.queue import Queue
 
-from .common.signals import ShutdownException
-from .core import dockerclient as dc
+from compose.common.signals import ShutdownException
+from compose.core import dockerclient as dc
 from compose.utils import get_output_stream
 
 log = logging.getLogger(__name__)
